requirements.txt
ollama
streamlit
python-dotenv
PyPDF2
